requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2026.1.4
greenlet==3.3.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
python-telegram-bot==13.15
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.46
tornado==6.1
typing_extensions==4.15.0
tzlocal==5.3.1
